refactor porting
