+++ pkg/gotrepo/staging.go
@@ -2,22 +2,16 @@ package gotrepo
 
 import (
 	"context"
-	"encoding/binary"
-	"encoding/json"
 	"os"
 
-	"github.com/brendoncarroll/go-state"
 	"github.com/brendoncarroll/go-state/posixfs"
 	"github.com/gotvc/got/pkg/branches"
-	"github.com/gotvc/got/pkg/gdat"
 	"github.com/gotvc/got/pkg/gotfs"
 	"github.com/gotvc/got/pkg/gotvc"
-	"github.com/gotvc/got/pkg/porting"
 	"github.com/gotvc/got/pkg/staging"
 	"github.com/gotvc/got/pkg/stores"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
-	bolt "go.etcd.io/bbolt"
 )
 
 // Add adds paths from the working directory to the staging area.
@@ -252,101 +246,7 @@ func (r *Repo) ForEachUntracked(ctx context.Context, fn func(p string) error) er
 	})
 }
 
-func (r *Repo) GetImportStores(ctx context.Context, branchName string) (*branches.StoreTriple, error) {
-	b, err := r.GetBranch(ctx, branchName)
-	if err != nil {
-		return nil, err
-	}
-	return r.getImportTriple(ctx, b)
-}
-
 func (r *Repo) getStage() *staging.Stage {
 	storage := newBoltKVStore(r.db, bucketStaging)
 	return staging.New(storage)
-}
-
-func (r *Repo) getImporter(ctx context.Context, b *branches.Branch) (*porting.Importer, error) {
-	salt := saltFromBytes(b.Salt)
-	saltHash := gdat.Hash(salt[:])
-	st, err := r.getImportTriple(ctx, b)
-	if err != nil {
-		return nil, err
-	}
-	fsop := r.getFSOp(b)
-	cache := portingCache{db: r.db, saltHash: saltHash}
-	return porting.NewImporter(fsop, cache, st.FS, st.Raw), nil
-}
-
-func (r *Repo) getExporter(b *branches.Branch) *porting.Exporter {
-	fsop := r.getFSOp(b)
-	cache := portingCache{db: r.db}
-	return porting.NewExporter(fsop, cache, r.workingDir)
-}
-
-func (r *Repo) getImportTriple(ctx context.Context, b *branches.Branch) (ret *branches.StoreTriple, _ error) {
-	salt := saltFromBytes(b.Salt)
-	saltHash := gdat.Hash(salt[:])
-	ids := [3]uint64{}
-	err := r.db.Update(func(tx *bolt.Tx) error {
-		ids = [3]uint64{}
-		b, err := tx.CreateBucketIfNotExists([]byte(bucketImportStores))
-		if err != nil {
-			return err
-		}
-		v := b.Get(saltHash[:])
-		if v == nil {
-			v = make([]byte, 8*3)
-			for i := 0; i < 3; i++ {
-				// TODO: maybe don't do this in a transaction
-				id, err := r.storeManager.Create(ctx)
-				if err != nil {
-					return err
-				}
-				binary.BigEndian.PutUint64(v[8*i:], id)
-			}
-			if err := b.Put(saltHash[:], v); err != nil {
-				return err
-			}
-		}
-		if len(v) != 8*3 {
-			return errors.New("bad length for staging store triple")
-		}
-		for i := 0; i < 3; i++ {
-			ids[i] = binary.BigEndian.Uint64(v[8*i:])
-		}
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-	return &branches.StoreTriple{
-		Raw: r.storeManager.Open(ids[0]),
-		FS:  r.storeManager.Open(ids[1]),
-		VC:  r.storeManager.Open(ids[2]),
-	}, nil
-}
-
-type portingCache struct {
-	db       *bolt.DB
-	saltHash [32]byte
-}
-
-func (c portingCache) Get(ctx context.Context, p string) (porting.Entry, error) {
-	return porting.Entry{}, state.ErrNotFound
-}
-
-func (c portingCache) Put(ctx context.Context, p string, ent porting.Entry) error {
-	_, err := json.Marshal(ent)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (c portingCache) Delete(ctx context.Context, p string) error {
-	return nil
-}
-
-func (c portingCache) List(ctx context.Context, span state.Span[string], ks []string) (int, error) {
-	return 0, nil
 }
